[refactor] move claim actions out of a nested folder.
